requirements.txt
flask
python-dotenv
spacy
scikit-learn
requests
numpy
